Rename src to lib
