+++ cli/lib/build.js
