+++ cli/lib/config.js
