+++ cli/lib/manifest.js
