+++ cli/lib/packages.js
