+++ cli/lib/registry.js
