+++ cli/lib/run.js
