+++ cli/lib/utils/each-object.js
