+++ cli/lib/utils/is.js
