+++ cli/lib/utils/map-object.js
