+++ cli/lib/utils/plural.js
