+++ cli/lib/utils/zip.js
