+++ cli/package.json
@@ -15,7 +15,7 @@
   },
   "bin": "bin/vba-blocks",
   "scripts": {
-    "format": "prettier --single-quote --write \"{bin,src,test}/**/*.{js,json}\"",
+    "format": "prettier --single-quote --write \"{bin,lib,test}/**/*.{js,json}\"",
     "pkg": "pkg --targets node6-win-x86,node6-win-x64,node6-macos-x64 --out-dir build ."
   },
   "dependencies": {
